change to float

+++ maritime_schema/maritime_types.py
@@ -517,7 +517,7 @@ class CagaEvent(BaseModel):
     route: List[Waypoint] = Field(
         None, description="Planned CAGA Route", examples=[waypoint_list_example])
 
-    calculation_time: Optional[int] = Field(
+    calculation_time: Optional[float] = Field(
         None, description="Time to calculate new route")
 
     class Config:
